+++ wallet/src/signer/trezor_signer/tests.rs
@@ -110,17 +110,21 @@ async fn test_sign_transaction_intent(#[case] seed: Seed) {
 }
 
 #[rstest]
-#[case(Seed::from_entropy(), SighashInputCommitmentVersion::V0)]
-#[case(Seed::from_entropy(), SighashInputCommitmentVersion::V1)]
+#[case(Seed::from_entropy(), true, SighashInputCommitmentVersion::V0)]
+#[case(Seed::from_entropy(), false, SighashInputCommitmentVersion::V1)]
 #[trace]
 #[serial]
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
 async fn test_sign_transaction(
     #[case] seed: Seed,
-    #[case] input_commitments_version: SighashInputCommitmentVersion,
+    #[case] before_fork: bool,
+    #[case] expected_input_commitments_version: SighashInputCommitmentVersion,
 ) {
     log::debug!(
-        "test_sign_transaction, seed = {seed:?}, input_commitments_version = {input_commitments_version:?}"
+        "test_sign_transaction, seed = {:?}, before_fork = {}, expected_input_commitments_version = {:?}",
+        seed,
+        before_fork,
+        expected_input_commitments_version
     );
 
     let _join_guard = maybe_spawn_auto_confirmer();
@@ -129,7 +133,8 @@ async fn test_sign_transaction(
 
     test_sign_transaction_generic(
         &mut rng,
-        input_commitments_version,
+        before_fork,
+        expected_input_commitments_version,
         make_trezor_signer,
         no_another_signer(),
         true,
@@ -256,17 +261,21 @@ async fn test_sign_transaction_intent_sig_consistency(#[case] seed: Seed) {
 }
 
 #[rstest]
-#[case(Seed::from_entropy(), SighashInputCommitmentVersion::V0)]
-#[case(Seed::from_entropy(), SighashInputCommitmentVersion::V1)]
+#[case(Seed::from_entropy(), true, SighashInputCommitmentVersion::V0)]
+#[case(Seed::from_entropy(), false, SighashInputCommitmentVersion::V1)]
 #[trace]
 #[serial]
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
 async fn test_sign_transaction_sig_consistency(
     #[case] seed: Seed,
-    #[case] input_commitments_version: SighashInputCommitmentVersion,
+    #[case] before_fork: bool,
+    #[case] expected_input_commitments_version: SighashInputCommitmentVersion,
 ) {
     log::debug!(
-        "test_sign_transaction_sig_consistency, seed = {seed:?}, input_commitments_version = {input_commitments_version:?}"
+        "test_sign_transaction_sig_consistency, seed = {:?}, before_fork = {}, expected_input_commitments_version = {:?}",
+        seed,
+        before_fork,
+        expected_input_commitments_version
     );
 
     let _join_guard = maybe_spawn_auto_confirmer();
@@ -275,7 +284,8 @@ async fn test_sign_transaction_sig_consistency(
 
     test_sign_transaction_generic(
         &mut rng,
-        input_commitments_version,
+        before_fork,
+        expected_input_commitments_version,
         make_deterministic_trezor_signer,
         Some(make_deterministic_software_signer),
         true,