+++ CHANGELOG.md
@@ -43,9 +43,16 @@ The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/
       of a cluster that exceeds the limits. The default cluster limits are: max 64 transactions, max 100'000 bytes
       in total.
 
+## [1.3.1] - 2026-06-03
+
 ### Fixed
   - Wallet:
-    - Fixed handling of confirmed and unconfirmed conflicting order transactions in the wallet.
+    - Fixed wallet handling of V1 order command conflicts, allowing confirmed ConcludeOrder and FreezeOrder transactions
+      to replace conflicting unconfirmed order transactions correctly.
+    - Fixed an issue where the wallet in cold mode would always use input commitments v0, thus producing signatures
+      that may no longer be valid at the current height.
+    - Fixed an issue where the wallet in cold mode would still try to access the mempool, which would cause `wallet-cli`
+      to print lots of error messages like "Method is not available in cold wallet mode".
 
 ## [1.3.0] - 2026-04-09